+++ data/pigui/modules/station-view/06-shipRepairs.lua
@@ -6,7 +6,7 @@ local StationView = require 'pigui.views.station-view'
 local ShipDef = require "ShipDef"
 local Game = require "Game"
 local Rand = require "Rand"
-local PiGuiFace = require 'ui/PiguiFace'
+local PiGuiFace = require 'pigui.libs.face'
 local Format = require "Format"
 local Character = require "Character"
 local ModalWindow = require 'pigui.libs.modal-win'
