+++ data/pigui/modules/station-view/07-police.lua
@@ -8,7 +8,7 @@ local StationView = require 'pigui.views.station-view'
 local Lang = require 'Lang'
 local Legal = require "Legal"
 local utils = require "utils"
-local PiGuiFace = require 'ui/PiguiFace'
+local PiGuiFace = require 'pigui.libs.face'
 local Format = require "Format"
 local Character = require "Character"
 local l = Lang.GetResource("ui-core")
